src/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-storages==1.14.2
Pillow==10.1.0
openai==1.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
django-redis==5.4.0
boto3==1.28.57
gunicorn==21.2.0
drf-yasg==1.21.7
whitenoise==6.5.0